Changing encoding of c strings returned for the linker on OS X to ASCII: I doubt the Mach-O file format is in UTF-8 Fixing releasing when the parent's class init fails: The proper way is to pass null when that happens, not release then pass null. Using fast enumeration where possible. Modified the bin/cue handling to also find the bin for .toc files. Note that there isn't a UTI for .toc files yet.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@79980 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ macosx/PluginController.m
@@ -20,7 +20,7 @@
 - (IBAction)selectPlugin:(id)sender
 {
 	if (sender==pluginMenu) {
-		int index = [pluginMenu indexOfSelectedItem];
+		NSInteger index = [pluginMenu indexOfSelectedItem];
 		if (index != -1) {
 			PcsxrPlugin *plugin = [plugins objectAtIndex:index];
